pyproject.toml
[build-system]
requires = ["setuptools>=61.0","wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-qa"
version = "0.1.0"
description = "AI 智能问答系统"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-core>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0"
]

[tool.setuptools.packages.find]
where = ["src"]